refactor(interpreter): extract setup_proxy into builtins/proxy.rs (#25)
Pure move: relocates the Proxy constructor + Proxy.revocable setup out
of the ~8.3k-line builtins/mod.rs into a new sibling file, following
the same pattern PR #163 used for the iterator-protocol cluster. The
function body is byte-identical; only its `impl Interpreter` wrapper
and visibility (now pub(crate), called back from setup_globals) differ.

First slice of a staged builtins/mod.rs split (part of #25).
